requirements.txt
# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0

# Async processing
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0

# Document processing and OCR
PyMuPDF==1.23.8
python-pptx==0.6.23
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3

# Enhanced OCR
easyocr==1.7.0
opencv-python==4.8.1.78

# FIXED: Compatible ML stack
torch==2.3.1
transformers==4.35.2
sentence-transformers==2.2.2
safetensors==0.4.0

# Phase 2: Hybrid search
rank-bm25==0.2.2

# Vector database
faiss-cpu==1.7.4

# Numerical computing
numpy==1.24.3
pandas==2.1.4

# HTTP and API
requests==2.31.0
httpx==0.28.1

# Environment and config
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Caching
diskcache==5.6.3

# Testing framework
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Google Generative AI
google-generativeai==0.3.2

# Rate limiting
flask-limiter==3.5.0

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0
